requirements.txt
Pillow==9.2.0
pytube==12.1.0
